+++ test/color_models_test.dart
@@ -144,7 +144,7 @@ void main() {
                   : _toColorModel(colorModel1, color2).toListWithAlpha();
 
               for (var steps = 1; steps <= 100; steps++) {
-                final colors = color1.interpolateTo(color2, steps);
+                final colors = color1.lerpTo(color2, steps);
 
                 for (var k = 0; k < colors.length; k++) {
                   final step = (1 / (steps + 1)) * k;